Issue - unsure how to convert IEnumerable object to List- resolved temporarily, commented out code.
The issue is causing a crash when when there are more than 1 non-formation (Bodyguard formations 3, 4 or 5). Another issue is with the number of captains. If there is only 1 captain, then the bodyguard formation is correctly assigned to follow the captain. However, every additional captain makes the formation follow the newest captain.

This causes the crash, if there are more than 1 captain.

I know that it is a Talesworld.Mountandblade.Agent object, but am unsure on how to convert it to a List form.