Update main.py
note some issues, fixed?

main issue - move movements is more sensitive, so mouse movements divded by 7 seems to work.

issue 2 - after the code runs, the target/final position of the mouse w/pyautogui or pydirectinput is different than the actual position of the mouse seen on the screen. When pyautogui is run again, it causes the screen to flip erractically.

Currently can track the waypoint and move towards it, but still need some fixes and to improve the image recognition. w/yolov8 & opencv.